+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export { default as pubSub } from './pub-sub';
 export * from './accordion';
+export * from './tabs';
 export * from './snackbar';