+++ src/features/games-improvements/achievements/default-status.ts
@@ -7,12 +7,16 @@ export const changeToDefaultStatus = async(): Promise<void> => {
   const url = getUrlProperties(window.location.href, ['pathname','search']);
   const status = await waitForElement(`#${gameAchievements.gameAchievementsDefaultStatusValue}`) as HTMLSelectElement;
 
-  if (url !== Cache.gameAchievementsDefaultStatusPathName) {
+  if (status && url !== Cache.gameAchievementsDefaultStatusPathName) {
     Cache.gameAchievementsDefaultStatusPathName = getUrlProperties(window.location.href, ['pathname','search']);
 
     if (GamesRegex.Test.achievementsUrlWithGamerId() && new URLSearchParams(window.location.search).get('gamerid') !== Constants.gamerId) return;
     if (!status.hasAttribute('checked')) {
-      status.click();
+      const gamerProgress = document.querySelector('.game.gamer .progress .progress-ring text');
+      
+      if (gamerProgress && parseInt(gamerProgress.textContent) !== 100) {
+        status.click();
+      }
     }
   }
 };