+++ src/types/components/tabs.ts
@@ -1,0 +1,5 @@
+export type TabTemplateOptions = {
+  link?: string,
+  id?: string,
+  content?: string,
+}