+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './features';
+export * from './components';
 export * from './memoizied-fetch-opts';
 export * from './sticky-element-opts';
 export * from './template-opts';